+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 1.0.5 
 
 * Feat: added `buildChildrenAsync` and related helper methods to create async rendering for Nodes (like github does).
 * Feat: added `index` property to `ComponentContext`.